InitialFileBeforeCommit.java
To take the help of these tests as a part of Automated Integrated Testing demonstration for my other Subject I wanted to add few more test cases and remove the comments so that the demonstration could be done better, but by mistake instead of making changes to the cloned repository I made the changes to my this file. Updated the source code again with all the previous visible comments and changes similar to the initially uploaded file, all the 24 tests are working as demonstrated in the video uploaded before the deadline.